cleanup(renderer): extract app bootstrap hooks
